requirements.txt
anthropic>=0.40.0
praw>=7.8.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
instagrapi>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
